pyproject.toml
[project]
name = "bruker-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "dmarek03",email = "domarek@student.agh.edu.pl"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "six (>=1.17.0,<2.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "crispy-bootstrap5 (>=2025.4,<2026.0)",
    "djangorestframework (>=3.15.0)",
    "drf-spectacular (>=0.27.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
